Version 3.27.30 (based on bleeding_edge revision r21851)
Fix representation of Phis for mutable-heapnumber-in-object-literal properties (issue 3392).

Performance and stability improvements on all platforms.

git-svn-id: https://v8.googlecode.com/svn/trunk@21852 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/heap.cc
@@ -4936,7 +4936,7 @@ void Heap::IterateStrongRoots(ObjectVisitor* v, VisitMode mode) {
 bool Heap::ConfigureHeap(int max_semi_space_size,
                          int max_old_space_size,
                          int max_executable_size,
-                         int code_range_size) {
+                         size_t code_range_size) {
   if (HasBeenSetUp()) return false;
 
   // Overwrite default configuration.
